feature(cli): #1694 shared response serialization util
