+++ packages/cli/src/lib/api-route-worker.js
@@ -1,33 +1,6 @@
 // https://github.com/nodejs/modules/issues/307#issuecomment-858729422
 import { parentPort } from "node:worker_threads";
-import { transformKoaRequestIntoStandardRequest } from "./resource-utils.js";
-
-// based on https://stackoverflow.com/questions/57447685/how-can-i-convert-a-request-object-into-a-stringifiable-object-in-javascript
-async function responseAsObject(response) {
-  if (!(response instanceof Response)) {
-    throw Object.assign(new Error(), {
-      name: "TypeError",
-      message: "Argument must be a Response object",
-    });
-  }
-  response = response.clone();
-
-  function stringifiableObject(obj) {
-    const filtered = {};
-    for (const key in obj) {
-      if (["boolean", "number", "string"].includes(typeof obj[key]) || obj[key] === null) {
-        filtered[key] = obj[key];
-      }
-    }
-    return filtered;
-  }
-
-  return {
-    ...stringifiableObject(response),
-    headers: Object.fromEntries(response.headers),
-    body: await response.text(),
-  };
-}
+import { transformKoaRequestIntoStandardRequest, responseAsObject } from "./resource-utils.js";
 
 async function executeRouteModule({ href, request, params }) {
   const { body, headers = {}, method, url } = request;
